Zoomable Treemap

Implementation of a Mike Bostock's d3 zoomable treemap as an htmlwidget in d3v5.

zoomable_treemap(data = jsonlite::toJSON(jsonlite::fromJSON(system.file("examples/flare.json",
  package = "d3RZoomableTreemap"))), width = "100%", height = NULL,
  elementId = NULL, background = "#bbb",
  header_background = "orange", header_height = 25,
  header_fontsize = "12px", format_string = ",",
  zoom_in_helptext = " - Click on a Square to Zoom",
  zoom_out_helptext = " - Click here to Zoom Out",
  tooltip_background = "orange", rmarkdown = FALSE)

Arguments

data
A json object

width
Width of the widget, in pixels

height
Height of the widget, in pixels

elementId
Element ID used for R Shiny package

background
Default background colour use if there is no color variable defined in the data. Takes hex of colour names.

header_background
The background colour either named of in a hex e.g. "#bbb"

header_height
Height of the title, in pixels

header_fontsize
Font size of the title, in pixels

format_string
D3 number format

zoom_in_helptext
The help text displayed on the header bar

zoom_out_helptext
Zoom out helptext

tooltip_background
Background colour of the tooltip

rmarkdown
Specific whether is an Rmarkdown document, to correct tooltip location.

Examples

data.frame(Titanic) %>%
 dplyr::select(Class,Age,Survived,Sex,Freq) %>%
  d3_nest2(value_col="Freq", root="Titanic",
           id_vars = c("Class", "Age", "Survived", "Sex")) %>%
  zoomable_treemap()